pyproject.toml
[tool.poetry]
name = "my-test-repo"
version = "0.1.0"
description = ""
authors = ["Amit Hilf <amithilf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygithub = "^2.3.0"
pytest-mock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
poetry = "^1.8.3"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
